cli out of bin
